Moved Configuration trait into configuration package
