+++ src/main/scala/play/api/cache/redis/package.scala
@@ -8,5 +8,6 @@ package object redis extends AnyRef with util.Expiration {
   type SynchronousResult[ A ]  = A
   type AsynchronousResult[ A ] = scala.concurrent.Future[ A ]
 
+  type Configuration = configuration.RedisConfiguration
   type RedisConnector = connector.RedisConnector
 }
